requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
ape-safe==0.7.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
base58==2.1.1
bitarray==2.6.0
black==22.10.0
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
construct==2.10.68
construct-classes==0.1.2
cytoolz==0.12.0
dataclassy==0.11.1
ecdsa==0.18.0
eip712==0.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-bloom==1.0.4
eth-brownie==1.19.2
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
frozenlist==1.3.1
hexbytes==0.2.3
hypothesis==6.27.3
idna==3.4
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
libusb1==3.0.0
lru-dict==1.1.8
mnemonic==0.20
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
py==1.11.0
py-ecc==5.2.0
py-evm==0.5.0a3
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycodestyle==2.10.0
pycryptodome==3.15.0
pyethash==0.1.27
Pygments==2.13.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
pysha3==1.0.2
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
PyYAML==5.4.1
requests==2.28.1
rlp==2.0.1
safe-eth-py==5.0.0
semantic-version==2.10.0
simple-rlp==0.1.3
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
trezor==0.13.5
trie==2.0.0a5
typing-extensions==3.10.0.2
urllib3==1.26.12
varint==1.0.2
vvm==0.1.0
vyper==0.3.7
wcwidth==0.2.5
web3==5.31.1
websockets==9.1
wrapt==1.14.1
yarl==1.8.1
